Pylint: Move its config page to its own module
